Split resistance out from typing
