+++ resistance/config.ini
@@ -1,0 +1,7 @@
+[scripts]
+script-path = /home/dfornika/code/cpo-pipeline/typing
+
+[databases]
+card = /data/ref_databases/card/card-2.0.3.json
+abricate-datadir = /projects/carbapenemase_producing_organisms_surveillance/databases/abricate
+abricate-cpo-plasmid-db = carbapenemases
