+++ resistance/job_scripts/abricate.sh
