+++ resistance/job_scripts/rgi.sh
