+++ resistance/test/data/pipeline_output/SAMPLE-ID/resistance/abricate/abricate.tsv
@@ -1,0 +1,2 @@
+#FILE	SEQUENCE	START	END	GENE	COVERAGE	COVERAGE_MAP	GAPS	%COVERAGE	%IDENTITY	DATABASE	ACCESSION	PRODUCT
+contigs.fa	contig00044	3183	3995	NDM-1	1-813/813	===============	0/0	100.00	100.00	bccdc	CAZ39946.1	  subclass B1 metallo-beta-lactamase NDM-1
