+++ resistance/test/data/pipeline_output/SAMPLE-ID/resistance/rgi/rgi.txt
@@ -1,0 +1,2 @@
+ORF_ID	Contig	Start	Stop	Orientation	Cut_Off	Pass_Bitscore	Best_Hit_Bitscore	Best_Hit_ARO	Best_Identities	ARO	Model_type	SNPs_in_Best_Hit_ARO	Other_SNPs	Drug Class	Resistance Mechanism	AMR Gene Family	Predicted_DNA	Predicted_Protein	CARD_Protein_Sequence	Percentage Length of Reference Sequence	ID	Model_ID
+contig00007_44 # 46711 # 49833 # -1 # ID=7_44;partial=00;start_type=ATG;rbs_motif=AGGAG;rbs_spacer=5-10bp;gc_cont=0.547	contig00007_44	46711	49833	-	Strict	1800	1894.78	mdtB	92.79	3000793	protein homolog model	n/a	n/a	aminocoumarin antibiotic	antibiotic efflux	resistance-nodulation-cell division (RND) antibiotic efflux pump	ATGCAGGTGTTACCTCCTGACAACACAGGCGGACCATCGCGCCTGTTCATTCTGCGCCCCGTCGCCACTACGCTGCTGATGGTCGCGATTTTGCTTGCCGGTATTATCGGTTATCGCTTTCTGCCAATCGCCGCCTTACCTGAAGTGGATTACCCCACGATTCAGGTCGTGACGCTCTACCCTGGCGCCAGCCCGGACGTCATGACCTCGGCGGTAACCGCGCCGCTTGAACGTCAGTTTGGCCAGATGTCTGGCTTAAAACAGATGTCATCACAAAGCTCCGGTGGCGCATCGGTCGTTACGTTACAGTTTCAACTGACGCTCCCGCTGGATGTTGCCGAGCAGGAAGTTCAGGCCGCGATTAACGCCGCCACCAATTTATTACCTAGCGATCTACCTAACCCTCCGGTCTACAGCAAGGTAAACCCGGCAGATCCGCCGATCATGACGCTGGCGGTCACTTCCAGCGCCATGCCGATGACGCAGGTGGAAGATATGGTCGAAACCCGCGTCGCGCAGAAAATCTCTCAGGTATCCGGCGTGGGTCTGGTAACACTTTCCGGTGGGCAACGTCCGGCGGTCCGCGTGAAACTTAACGCTCAGGCCATTGCCGCACTCGGTCTGACCAGCGAAACCATCCGTACGGCCATTACCGGTGCCAACGTCAATTCAGCAAAAGGGAGTCTTGATGGCCCAACGCGCGCCGTTACCCTTTCAGCTAACGACCAGATGCAGTCCGCAGATGAATATCGTCAGCTCATTATTGCCTATAAAAACGGTGCACCGGTTCGCCTGGGCGATGTCGCTACCGTCGAACAAGGAGCAGAAAACAGCTGGCTGGGTGCATGGGCAAATAAGCAACAAGCTATAGTGATGAATGTGCAGCGCCAGCCTGGGGCGAACATCATCGCCACGGCAGACAGTATCCGCCAGATGCTGCCACAGCTCACCGAAAGCCTGCCCAAATCGGTTAAGGTCAGCGTACTTTCCGATCGCACCACCAATATTCGTGCCTCTGTTGCAGACACCCAGTTCGAGTTGATGCTGGCGATAGCGCTGGTGGTGATGATTATCTACCTGTTTCTGCGCAATGTTCCGGCGACGATTATTCCTGCAGTCGCCGTACCGTTGTCGTTAATCGGCACCTTCGCGGTGATGGTATTTCTTGATTTCTCCATCAACAACCTGACCTTAATGGCGCTAACCATCGCGACAGGCTTTGTTGTGGATGACGCTATCGTGGTTATCGAGAATATTTCACGCTATATCGAAAAAGGCGAAAAACCGCTGACTGCTGCCCTGAAGGGCGCAGGCGAGATTGGTTTTACCATCATCTCTCTGACGTTTTCATTGATTGCGGTATTGATCCCACTGCTATTCATGGGCGATATCGTCGGGCGCTTGTTCCGTGAATTCGCCGTTACGCTGGCGGTGGCGATTCTGATTTCCGCCGTTGTCTCGCTCACTCTTACGCCGATGATGTGCGCGCGTATGCTGAATCAGGCCTCATTGCGTAAGCAGAATCGCTTCTCCCGCGCCTCAGAGCGAGTGTTTGATCGCGTCATAGCCGGGTACGGTCATTGGCTGGCAAAAGTACTTAACCACCCGTGGCTGACGCTGAGCGTAGCGCTCGGCACGTTGGCATTGAGCGTCATGCTATGGGTTTTTATCCCCAAAGGTTTCTTCCCAATCCAGGACAACGGCATTATCCAGGGTACGCTACAGGCACCGCAGTCCAGCTCATTCGCCAGCATGGCCGAGCGCCAGCGTCAGGTTTCAGATGTCATCCTTAAAGATCCGGCGGTAGAGAGCCTGACCGCTTTTGTCGGCGTGGATGGTACTAATCCGGCATTGAACAGCGCACGTTTGCAAATCAACCTTAAACCGCTGGATGAACGTGACGATCGGGTGCAGAAAGTGATTGCGCGCTTACAGGACGCGGTTGACAGGGTTCCCGGAGTTGACCTTTATCTTCAGCCTACTCAGGATCTGACGATTGATACTCAGGTGAGCCGGACGCAATATCAGTTCACGCTGCAAGCGACATCCCTTGATGCCCTCAGCACCTGGGTTCCACAGCTGATAGCAAAACTAAATCAGCTACCGCAGCTGTCCGATGTCAGCAGCGACTGGCAGGACAAAGGGCTGGTAGCGTATGTGAACGTCAATCGCGACAGCGCCAGCCGTTTGGGCATCAGCATGTCAGATGTCGACAACGCCCTGTACAACGCCTTCGGCCAGCGTTTGATTTCGACTATTTACACCCAGGCAAATCAGTATCGCGTGGTGTTAGAACATAACACCAACAATACGCCGGGGCTTGCGGCACTGGATACAATACGTCTGACCAGTATCGACGGTGGCATAGTACCGCTCAGCGCGATCGCCAGTATTGAACAGCGTTTTGCCCCGCTGTCGATTAACCACCTCGATCAGTTCCCGGTCACCACCATCTCGTTCAACGTGCCAGACGAACATTCGCTGGGTGATGCGGTACAGGCGATTCTGGACGCCGAAAAAACGCTTAACCTGCCCGCAAATATTACGACCCAGTTCCAGGGTAGCACGCTCGCCTTCCAGGCCGCGCTGGGAAATACCATCTGGTTGATTGTGGCTGCGGTTGTCGCGATGTATATCGTCCTCGGCGTGCTGTATGAGAGCTTTATTCACCCAATTACTATTCTCTCCACGCTGCCGACGGCGGGCGTCGGGGCACTGCTGGCGTTAATGATTGCCGGCAGCGAACTGGACGTCATCGCGATTATTGGCATTATTTTGCTCATCGGTATCGTCAAGAAAAACGCCATAATGATGATCGACTTTGCCCTTGCCGCGGAGCGCGAACAAGGCATGGCGCCGCGCGATGCTATTTTCCAGGCCTGTCTGCTGCGTTTCCGCCCAATTCTGATGACTACGCTGGCGGCACTGCTCGGCGCGTTGCCGCTGATGTTAAGTACCGGTGTGGGCGCAGAGTTACGCCGTCCGTTAGGCATTGGCATGGTCGGTGGACTGCTTGTGAGTCAGGTATTAACTCTGTTTACCACTCCGGTAATTTATCTGCTGTTTGACCGTCTGTCGCTGTATGTGAAAAGCCGTTTCCCGCGTCAGGAAGAGGAAGCGTAA	MQVLPPDNTGGPSRLFILRPVATTLLMVAILLAGIIGYRFLPIAALPEVDYPTIQVVTLYPGASPDVMTSAVTAPLERQFGQMSGLKQMSSQSSGGASVVTLQFQLTLPLDVAEQEVQAAINAATNLLPSDLPNPPVYSKVNPADPPIMTLAVTSSAMPMTQVEDMVETRVAQKISQVSGVGLVTLSGGQRPAVRVKLNAQAIAALGLTSETIRTAITGANVNSAKGSLDGPTRAVTLSANDQMQSADEYRQLIIAYKNGAPVRLGDVATVEQGAENSWLGAWANKQQAIVMNVQRQPGANIIATADSIRQMLPQLTESLPKSVKVSVLSDRTTNIRASVADTQFELMLAIALVVMIIYLFLRNVPATIIPAVAVPLSLIGTFAVMVFLDFSINNLTLMALTIATGFVVDDAIVVIENISRYIEKGEKPLTAALKGAGEIGFTIISLTFSLIAVLIPLLFMGDIVGRLFREFAVTLAVAILISAVVSLTLTPMMCARMLNQASLRKQNRFSRASERVFDRVIAGYGHWLAKVLNHPWLTLSVALGTLALSVMLWVFIPKGFFPIQDNGIIQGTLQAPQSSSFASMAERQRQVSDVILKDPAVESLTAFVGVDGTNPALNSARLQINLKPLDERDDRVQKVIARLQDAVDRVPGVDLYLQPTQDLTIDTQVSRTQYQFTLQATSLDALSTWVPQLIAKLNQLPQLSDVSSDWQDKGLVAYVNVNRDSASRLGISMSDVDNALYNAFGQRLISTIYTQANQYRVVLEHNTNNTPGLAALDTIRLTSIDGGIVPLSAIASIEQRFAPLSINHLDQFPVTTISFNVPDEHSLGDAVQAILDAEKTLNLPANITTQFQGSTLAFQAALGNTIWLIVAAVVAMYIVLGVLYESFIHPITILSTLPTAGVGALLALMIAGSELDVIAIIGIILLIGIVKKNAIMMIDFALAAEREQGMAPRDAIFQACLLRFRPILMTTLAALLGALPLMLSTGVGAELRRPLGIGMVGGLLVSQVLTLFTTPVIYLLFDRLSLYVKSRFPRQEEEA	MQVLPPSSTGGPSRLFIMRPVATTLLMVAILLAGIIGYRALPVSALPEVDYPTIQVVTLYPGASPDVMTSAVTAPLERQFGQMSGLKQMSSQSSGGASVITLQFQLTLPLDVAEQEVQAAINAATNLLPSDLPNPPVYSKVNPADPPIMTLAVTSTAMPMTQVEDMVETRVAQKISQISGVGLVTLSGGQRPAVRVKLNAQAIAALGLTSETVRTAITGANVNSAKGSLDGPSRAVTLSANDQMQSAEEYRQLIIAYQNGAPIRLGDVATVEQGAENSWLGAWANKEQAIVMNVQRQPGANIISTADSIRQMLPQLTESLPKSVKVTVLSDRTTNIRASVDDTQFELMMAIALVVMIIYLFLRNIPATIIPGVAVPLSLIGTFAVMVFLDFSINNLTLMALTIATGFVVDDAIVVIENISRYIEKGEKPLAAALKGAGEIGFTIISLTFSLIAVLIPLLFMGDIVGRLFREFAITLAVAILISAVVSLTLTPMMCARMLSQESLRKQNRFSRASEKMFDRIIAAYGRGLAKVLNHPWLTLSVALSTLLLSVLLWVFIPKGFFPVQDNGIIQGTLQAPQSSSFANMAQRQRQVADVILQDPAVQSLTSFVGVDGTNPSLNSARLQINLKPLDERDDRVQKVIARLQTAVDKVPGVDLFLQPTQDLTIDTQVSRTQYQFTLQATSLDALSTWVPQLMEKLQQLPQLSDVSSDWQDKGLVAYVNVDRDSASRLGISMADVDNALYNAFGQRLISTIYTQANQYRVVLEHNTENTPGLAALDTIRLTSSDGGVVPLSSIAKIEQRFAPLSINHLDQFPVTTISFNVPDNYSLGDAVQAIMDTEKTLNLPVDITTQFQGSTLAFQSALGSTVWLIVAAVVAMYIVLGILYESFIHPITILSTLPTAGVGALLALLIAGSELDVIAIIGIILLIGIVKKNAIMMIDFALAAEREQGMSPREAIYQACLLRFRPILMTTLAALLGALPLMLSTGVGAELRRPLGIGMVGGLIVSQVLTLFTTPVIYLLFDRLALWTKSRFARHEEEA	100.00	gnl|BL_ORD_ID|776|hsp_num:0	820
