+++ typing/config.ini
@@ -2,7 +2,4 @@
 script-path = /home/dfornika/code/cpo-pipeline/typing
 
 [databases]
-card = /data/ref_databases/card/card-2.0.3.json
-abricate-datadir = /projects/carbapenemase_producing_organisms_surveillance/databases/abricate
-abricate-cpo-plasmid-db = carbapenemases
 mlst-scheme-map = /home/dfornika/code/cpo-pipeline/typing/scheme_species_map.tab
