+++ typing/pipeline.py
@@ -1,10 +1,9 @@
 #!/usr/bin/env python
 
 '''
-This script is a wrapper for module two, part 1: typing and resistance gene idenfification from assemblies. 
+This script is a wrapper for module two, part 1: typing from assemblies. 
 
-It uses mlst, resfinder, plasmidfinder, rgi, mobsuite for sequence typing, AMR profile prediction and plasmid predictions. 
-The output is a TSV summary of the results. 
+It uses mlst, mobsuite for sequence typing. 
 
 Example usage:
 
@@ -54,9 +53,6 @@ def main():
 
     parser.add_option("-s", "--mlst-scheme-map", dest="mlst_scheme_map_file", default=config['databases']['mlst-scheme-map'], type="string", help="absolute file path to mlst scheme")
     parser.add_option("-k", "--script-path", dest="script_path", default=config['scripts']['script-path'], type="string", help="absolute file path to this script folder")
-    parser.add_option("-c", "--card-path", dest="card_path", default=config['databases']['card'], type="string", help="absolute file path to card.json db")
-    parser.add_option("-d", "--abricate-datadir", dest="abricate_datadir", default=config['databases']['abricate-datadir'], type="string", help="absolute file path to directory where abricate dbs are stored")
-    parser.add_option("-p", "--abricate-cpo-plasmid-db", dest="abricate_cpo_plasmid_db", default=config['databases']['abricate-cpo-plasmid-db'], type="string", help="name of abricate cpo plasmid db to use")
     
     
     (options, args) = parser.parse_args()
@@ -64,9 +60,6 @@ def main():
     ID = str(options.id).lstrip().rstrip()
     assembly = options.assembly
     script_path = options.script_path
-    card_path = options.card_path
-    abricate_datadir = options.abricate_datadir
-    abricate_cpo_plasmid_db = options.abricate_cpo_plasmid_db
     mlst_scheme_map_file = options.mlst_scheme_map_file
     outputDir = options.output
 
@@ -76,8 +69,6 @@ def main():
     file_paths = {
         'mlst_path': '/'.join([outputDir, ID, 'typing', 'mlst', 'mlst.tsv']),
         'mob_recon_path': '/'.join([outputDir, ID, 'typing', 'mob_recon']),
-        'abricate_path': '/'.join([outputDir, ID, 'resistance', 'abricate', 'abricate.tsv']),
-        'rgi_path': "/".join([outputDir, ID, 'resistance', 'rgi', 'rgi'])
     }
     
     job_script_path = os.path.join(os.path.dirname(os.path.realpath(sys.argv[0])),  'job_scripts')
@@ -108,41 +99,7 @@ def main():
         running_jobs = [session.runJob(job) for job in prepared_jobs]
         for job_id in running_jobs:
             print('Your job has been submitted with ID %s' % job_id)
-        session.synchronize(running_jobs, drmaa.Session.TIMEOUT_WAIT_FOREVER, True)
-    
-    resistance_jobs = [
-        {
-            'job_name': 'abricate',
-            'native_specification': '-pe smp 8',
-            'remote_command': os.path.join(job_script_path, 'abricate.sh'),
-            'args': [
-                "--input", assembly,
-                "--datadir", abricate_datadir,
-                "--database", abricate_cpo_plasmid_db,
-                "--output_file", file_paths['abricate_path']
-            ]
-        },
-        {
-            'job_name': 'rgi',
-            'native_specification': '-pe smp 8',
-            'remote_command': os.path.join(job_script_path, 'rgi.sh'),
-            'args': [
-                "--input", assembly,
-                "--card_json", card_path,
-                "--output_file", file_paths['rgi_path']
-            ]
-        }
-    ]
-    
-    with drmaa.Session() as session:
-        prepared_jobs = [prepare_job(job, session) for job in resistance_jobs]
-        running_jobs = [session.runJob(job) for job in prepared_jobs]
-        for job_id in running_jobs:
-            print('Your job has been submitted with ID %s' % job_id)
-        session.synchronize(running_jobs, drmaa.Session.TIMEOUT_WAIT_FOREVER, True)
-    
-    
-    print("step 3: parsing mlst, plasmid, and amr results")
+        session.synchronize(running_jobs, drmaa.Session.TIMEOUT_WAIT_FOREVER, True)    
     
     print("identifying MLST")
     mlst_report = "/".join([outputDir, ID, "typing", "mlst", "mlst.tsv"]) 
@@ -236,43 +193,7 @@ def main():
     plasmid_contigs = get_plasmid_contigs(mob_recon_contig_report)
     likely_plasmid_contigs = get_likely_plasmid_contigs(mob_recon_contig_report)
     origins = get_plasmid_origins(mob_recon_contig_report)
-    
-    abricate_report_path = "/".join([outputDir, ID, "resistance", "abricate", "abricate.tsv"])
-    abricate_report = result_parsers.parse_abricate_result(abricate_report_path)
-    
-    rgi_report_path = "/".join([outputDir, ID, "resistance", "rgi", "rgi.txt"])
-    rgi_report = result_parsers.parse_rgi_result_txt(rgi_report_path)
 
-    def get_abricate_carbapenemases(abricate_report):
-        """
-        Given a list of dicts generated by parsing an abricate report file, 
-        return a list of carbapenemases.
-        Args:
-            abricate_report (list of dict):
-        Returns:
-            list: likely plasmid contigs
-            For example: ['NDM-1', '', '', ...]
-        """
-        abricate_carbapenemases = []
-        for abricate_report_record in abricate_report:
-            abricate_carbapenemases.append(abricate_report_record['gene'])
-        return abricate_carbapenemases
-    
-    def get_rgi_carbapenemases(rgi_report):
-        """
-        Given a list of dicts generated by parsing an rgi report file, 
-        return a list of carbapenemases.
-        Args:
-            rgi_report (list of dict):
-        Returns:
-            list: likely plasmid contigs
-            For example: ['', '', '', ...]
-        """
-        rgi_carbapenemases = []
-        for rgi_report_record in rgi_report:
-            if re.search("carbapenem", rgi_report_record['drug_class']):
-                rgi_carbapenemases.append(rgi_report_record['best_hit_aro'])
-        return rgi_carbapenemases
 
 if __name__ == "__main__":
     start = time.time()
